Extract transaction related code to its own util file from types.js
