+++ src/components/ActivityFeed/ActivityFeed.example.js
@@ -20,7 +20,7 @@ import {
   TRANSITION_REVIEW_2_BY_PROVIDER,
   TX_TRANSITION_ACTOR_CUSTOMER,
   TX_TRANSITION_ACTOR_PROVIDER,
-} from '../../util/types';
+} from '../../util/transaction';
 import ActivityFeed from './ActivityFeed';
 
 const noop = () => null;
