+++ src/components/BookingBreakdown/BookingBreakdown.test.js
@@ -4,11 +4,11 @@ import { fakeIntl, createBooking } from '../../util/test-data';
 import { renderDeep } from '../../util/test-helpers';
 import { types as sdkTypes } from '../../util/sdkLoader';
 import {
-  LINE_ITEM_NIGHT,
   TRANSITION_CANCEL,
   TRANSITION_REQUEST,
   TX_TRANSITION_ACTOR_CUSTOMER,
-} from '../../util/types';
+} from '../../util/transaction';
+import { LINE_ITEM_NIGHT } from '../../util/types';
 import { BookingBreakdownComponent } from './BookingBreakdown';
 
 const { UUID, Money } = sdkTypes;
