+++ src/containers/TransactionPage/TransactionPage.duck.js
@@ -13,7 +13,7 @@ import {
   TRANSITION_REVIEW_1_BY_PROVIDER,
   TRANSITION_REVIEW_2_BY_CUSTOMER,
   TRANSITION_REVIEW_2_BY_PROVIDER,
-} from '../../util/types';
+} from '../../util/transaction';
 import * as log from '../../util/log';
 import {
   updatedEntities,
