+++ src/forms/BookingDatesForm/BookingDatesForm.test.js
@@ -4,7 +4,8 @@ import Decimal from 'decimal.js';
 import { types as sdkTypes } from '../../util/sdkLoader';
 import { renderShallow, renderDeep } from '../../util/test-helpers';
 import { fakeIntl } from '../../util/test-data';
-import { LINE_ITEM_NIGHT, TRANSITION_REQUEST } from '../../util/types';
+import { TRANSITION_REQUEST } from '../../util/transaction';
+import { LINE_ITEM_NIGHT } from '../../util/types';
 import { dateFromAPIToLocalNoon } from '../../util/dates';
 import { BookingBreakdown } from '../../components';
 import { BookingDatesFormComponent } from './BookingDatesForm';
